# contracts/openapi/careos-foundation.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ROOTOPATHY CareOS Foundation API",
    "version": "0.21.0",
    "description": "Checked contract for the currently implemented public foundation, browser identity with approved role-level MFA enforcement, approved-registry governed invitations, maker-checker MFA and membership administration, membership-backed organization selection and access projection, and bounded organization-core administration endpoints with the exact approved readiness, organization-profile, registration-identifier, effective address/contact, and versioned international-settings contracts. Unsafe browser operations require an approved Origin or Referer plus the double-submit CSRF cookie/header pair."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1/organizations/{organizationId}/facilities": {
      "get": { "operationId": "getFacilityDirectory", "parameters": [{ "$ref": "#/components/parameters/OrganizationId" }, { "name": "query", "in": "query", "schema": { "type": "string", "maxLength": 120 } }, { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "under_review", "active", "suspended", "closed"] } }], "security": [{ "sessionCookie": [] }], "responses": { "200": { "description": "Authorized facility directory", "headers": { "X-Correlation-Id": { "$ref": "#/components/headers/CorrelationId" }, "X-CareOS-Session-Expires-In": { "$ref": "#/components/headers/SessionExpiresIn" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FacilityDirectory" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "500": { "$ref": "#/components/responses/InternalError" } } },
      "post": { "operationId": "createFacilityDraft", "parameters": [{ "$ref": "#/components/parameters/OrganizationId" }, { "$ref": "#/components/parameters/BrowserOrigin" }, { "$ref": "#/components/parameters/IdempotencyKey" }], "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FacilityCreateRequest" } } } }, "responses": { "201": { "description": "Facility draft created", "headers": { "X-Correlation-Id": { "$ref": "#/components/headers/CorrelationId" }, "X-CareOS-Session-Expires-In": { "$ref": "#/components/headers/SessionExpiresIn" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FacilityDirectory" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "409": { "$ref": "#/components/responses/Conflict" }, "500": { "$ref": "#/components/responses/InternalError" } } }
    },
    "/api/v1/organizations/{organizationId}/facilities/{facilityId}": {
      "put": { "operationId": "updateFacilityDraft", "parameters": [{ "$ref": "#/components/parameters/OrganizationId" }, { "name": "facilityId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }, { "$ref": "#/components/parameters/BrowserOrigin" }, { "$ref": "#/components/parameters/IfMatch" }, { "$ref": "#/components/parameters/IdempotencyKey" }], "security": [{ "sessionCookie": [], "csrfCookie": [], "csrfHeader": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FacilityCreateRequest" } } } }, "responses": { "200": { "description": "Facility draft updated or exact idempotent response replayed", "headers": { "X-Correlation-Id": { "$ref": "#/components/headers/CorrelationId" }, "X-CareOS-Session-Expires-In": { "$ref": "#/components/headers/SessionExpiresIn" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FacilityDirectory" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "403": { "$ref": "#/components/responses/Forbidden" }, "404": { "$ref": "#/components/responses/NotFound" }, "409": { "$ref": "#/components/responses/Conflict" }, "412": { "$ref": "#/components/responses/PreconditionFailed" }, "428": { "$ref": "#/components/responses/PreconditionRequired" }, "500": { "$ref": "#/components/responses/InternalError" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "CAREOS_SESSION",
        "description": "Opaque HttpOnly SameSite=Strict server-side session identifier. Secure is mandatory outside explicit local HTTP development."
      },
      "csrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "XSRF-TOKEN",
        "description": "Readable SameSite=Strict double-submit token cookie. Secure is mandatory outside explicit local HTTP development."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN",
        "description": "Must match the raw XSRF-TOKEN cookie value."
      }
    },
    "parameters": {
      "BrowserOrigin": {
        "name": "Origin",
        "in": "header",
        "required": true,
        "description": "Must exactly match a configured CareOS browser origin. A same-origin Referer is accepted when Origin is unavailable.",
        "schema": {
          "type": "string",
          "format": "uri"
        }
      },
      "OrganizationId": {
        "name": "organizationId",
        "in": "path",
        "required": true,
        "description": "Organization boundary for every protected business-resource route.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "description": "Strong entity tag from the latest representation. Required for protected updates and deletes.",
        "schema": {
          "type": "string",
          "pattern": "^\"[A-Za-z0-9._:-]{1,128}\"$"
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Caller-generated key for an explicitly retryable protected mutation. Reuse is valid only for an equivalent request.",
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9._:-]+$"
        }
      }
    },
    "headers": {
      "CorrelationId": {
        "description": "Stable identifier used to correlate one HTTP request across logs and problem responses.",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9._:-]+$"
        }
      },
...
      "SessionExpiresIn": {
        "description": "Effective seconds remaining before the authenticated browser session reaches its idle or absolute deadline. Omitted when no authenticated session exists.",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "maximum": 2147483647
        }
      }
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Request validation, password policy, or one-time-token failure",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Credentials, verification evidence, or session is invalid",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Origin, CSRF, or authorization check failed",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "The resource is unavailable or hidden from the current actor",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "The requested transition, idempotency key, or current resource state conflicts with the operation",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "A required precondition is missing, such as recent authentication, recent MFA, or If-Match",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "PreconditionFailed": {
        "description": "The supplied entity tag no longer matches the current representation",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "The request failed without exposing sensitive implementation details.",
        "headers": {
          "X-Correlation-Id": {
            "$ref": "#/components/headers/CorrelationId"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "FacilityType": { "type": "object", "additionalProperties": false, "required": ["key", "displayName"], "properties": { "key": { "type": "string" }, "displayName": { "type": "string" } } },
      "FacilitySummary": { "type": "object", "additionalProperties": false, "required": ["facilityId", "facilityCode", "legalName", "displayName", "facilityType", "status", "lockVersion", "createdAt", "updatedAt"], "properties": { "facilityId": { "type": "string", "format": "uuid" }, "facilityCode": { "type": "string", "pattern": "^[A-Z0-9][A-Z0-9_-]{1,31}$" }, "legalName": { "type": "string" }, "displayName": { "type": "string" }, "facilityType": { "type": "string" }, "timezone": { "type": ["string", "null"] }, "status": { "type": "string", "enum": ["draft", "under_review", "active", "suspended", "closed"] }, "lockVersion": { "type": "integer", "format": "int64", "minimum": 0 }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" } } },
      "FacilityDirectory": { "type": "object", "additionalProperties": false, "required": ["organizationId", "canCreate", "facilityTypes", "facilities", "evaluatedAt"], "properties": { "organizationId": { "type": "string", "format": "uuid" }, "canCreate": { "type": "boolean" }, "facilityTypes": { "type": "array", "items": { "$ref": "#/components/schemas/FacilityType" } }, "facilities": { "type": "array", "items": { "$ref": "#/components/schemas/FacilitySummary" } }, "evaluatedAt": { "type": "string", "format": "date-time" } } },
      "FacilityCreateRequest": { "type": "object", "additionalProperties": false, "required": ["facilityCode", "legalName", "displayName", "facilityType", "reason"], "properties": { "facilityCode": { "type": "string", "pattern": "^[A-Z0-9][A-Z0-9_-]{1,31}$" }, "legalName": { "type": "string", "minLength": 2, "maxLength": 200 }, "displayName": { "type": "string", "minLength": 2, "maxLength": 120 }, "facilityType": { "type": "string" }, "addressId": { "type": ["string", "null"], "format": "uuid" }, "contactId": { "type": ["string", "null"], "format": "uuid" }, "timezone": { "type": ["string", "null"], "maxLength": 64 }, "reason": { "type": "string", "minLength": 10, "maxLength": 500 } } },
...
      "FieldViolation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["path", "code", "message"],
        "properties": {
          "path": {
            "type": "string",
            "pattern": "^/"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "correlationId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            }
          }
        }
      }
    }
  }
}